flask-app/requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch==1.11.0+cpu
# transformers==4.18.0

flask
data-preprocessors
bangla-postaggers

